--- a/src/plugins/discover/public/application/view_components/canvas/discover_table.tsx
+++ b/src/plugins/discover/public/application/view_components/canvas/discover_table.tsx
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import React, { useCallback, useMemo } from 'react';
+import React, { useCallback } from 'react';
 import { DiscoverViewServices } from '../../../build_services';
 import { useOpenSearchDashboards } from '../../../../../opensearch_dashboards_react/public';
 import { DataGridTable } from '../../components/data_grid/data_grid_table';
@@ -20,7 +20,6 @@ import {
 import { IndexPatternField, opensearchFilters } from '../../../../../data/public';
 import { DocViewFilterFn } from '../../doc_views/doc_views_types';
 import { SortOrder } from '../../../saved_searches/types';
-import { DOC_HIDE_TIME_COLUMN_SETTING } from '../../../../common';
 import { OpenSearchSearchHit } from '../../doc_views/doc_views_types';
 import { popularizeField } from '../../helpers/popularize_field';
 
@@ -32,7 +31,6 @@ interface Props {
 export const DiscoverTable = ({ rows, scrollToTop }: Props) => {
   const { services } = useOpenSearchDashboards<DiscoverViewServices>();
   const {
-    uiSettings,
     data: {
       query: { filterManager },
     },
@@ -85,10 +83,6 @@ export const DiscoverTable = ({ rows, scrollToTop }: Props) => {
     },
     [filterManager, indexPattern]
   );
-  const displayTimeColumn = useMemo(
-    () => !!(!uiSettings.get(DOC_HIDE_TIME_COLUMN_SETTING, false) && indexPattern?.isTimeBased()),
-    [indexPattern, uiSettings]
-  );
 
   if (indexPattern === undefined) {
     // TODO: handle better
@@ -112,7 +106,6 @@ export const DiscoverTable = ({ rows, scrollToTop }: Props) => {
       onSort={onSetSort}
       sort={sort}
       rows={rows}
-      displayTimeColumn={displayTimeColumn}
       title={savedSearch?.id ? savedSearch.title : ''}
       description={savedSearch?.id ? savedSearch.description : ''}
       scrollToTop={scrollToTop}