--- a/test/functional/apps/dashboard/dashboard_time_picker.js
+++ b/test/functional/apps/dashboard/dashboard_time_picker.js
@@ -67,15 +67,15 @@ export default function ({ getService, getPageObjects }) {
         name: 'saved search',
         fields: ['bytes', 'agent'],
       });
-      // Current data grid loads 100 rows per page by default with inspect button and time range
-      await dashboardExpect.savedSearchRowCountFromLegacyTable(100);
+      // DefaultDiscoverTable loads 10 rows initially
+      await dashboardExpect.rowCountFromDefaultDiscoverTable(10);
 
       // Set to time range with no data
       await PageObjects.timePicker.setAbsoluteRange(
         'Jan 1, 2000 @ 00:00:00.000',
         'Jan 1, 2000 @ 01:00:00.000'
       );
-      await dashboardExpect.savedSearchRowCountFromLegacyTable(0);
+      await dashboardExpect.rowCountFromDefaultDiscoverTable(0);
     });
 
     it('Timepicker start, end, interval values are set by url', async () => {