### Description

Removed the simultaneous rendering of `DefaultDiscoverTable` and `DataGrid` when only one would be displayed.
Bumped DOMPurify for a boost in performance.

### Issues Resolved

<!-- List any issues this PR will resolve. Prefix the issue with the keyword closes, fixes, fix -->
<!-- Example: closes #1234 or fixes <Issue_URL> -->

## Screenshot

#### 2.13 - 10s span 

* Click to start of network call: 2392ms
* Data to end of render: 5384ms 
* Click to end of render: 8016ms (network call 256ms)

![2.13](https://github.com/opensearch-project/OpenSearch-Dashboards/assets/3527403/f81ca88a-5325-4daf-9c29-39392cdc6fb4)

#### PR - 10s span

* Click to start of network call: 1289ms - 46% improvement
* Data to end of render: 1626ms - 70% improvement
* Click to end of render: 3186ms (network call 261ms) - 62% improvement

![PRed](https://github.com/opensearch-project/OpenSearch-Dashboards/assets/3527403/904eb2e6-24ec-47c0-96f1-243bd68f0218)


## Changelog
- feat: Improve the perceived performance of Discover when using the default tabular renderer

### Check List

- [ ] All tests pass
  - [ ] `yarn test:jest`
  - [ ] `yarn test:jest_integration`
- [ ] New functionality includes testing.
- [ ] New functionality has been documented.
- [ ] Update [CHANGELOG.md](./../CHANGELOG.md)
- [X] Commits are signed per the DCO using --signoff