@virajsanghvi 2024-04-24
One question regarding perf metrics: are those tracked somewhere? Just curious if there's a way of validating the expected gains hold in production builds.