@AMoo-Miki 2024-04-24
> are those tracked somewhere? 

No but I intend to add code coverage and performance metrics gathering to my test framework.

> Just curious if there's a way of validating the expected gains hold in production builds.

The graphs I placed above were from prod builds.